// lib/enums.ts
export enum IconsRoutes {
  BOOK = "icons/book",
  CLOCK = "icons/clock",
  CROSS = "icons/cross",
  CORRECT = "icons/correct",
  EMPTY = "icons/empty",
  FFMPEG = "icons/ffmpeg",
  FILL = "icons/fill",
  FREQUENCY = "icons/frequency",
  LOGIN = "icons/login",
  LOGOUT = "icons/logout",
  NEXT = "icons/next",
  NOTES = "icons/notes",
  PAUSE = "icons/pause",
  PHOTOS = "icons/photos",
  PLAY = "icons/play",
  PLUS = "icons/plus",
  PREVIOUS = "icons/previous",
  REFRESH = "icons/refresh",
  RIGHT = "icons/right",
  SETTINGS = "icons/settings",
  STAR = "icons/star",
  UP = "icons/up",
  VERIFIED = "icons/verified",
}

export enum Version {
  V1 = "v1",
}

export enum SfwRoutes {
  BITE = "sfw/bite",
  NEKO = "sfw/neko",
  POKE = "sfw/poke",
  RUN = "sfw/run",
  SLAP = "sfw/slap",
  SMILE = "sfw/smile",
}

export enum OutputType {
  RANDOM = "random",
  GIFONLY = "gifOnly",
  PICTUREONLY = "pictureOnly",
}

export enum DataStorageType {
  JSON = "json",
}